+++ src/ENDFtk/section/BaseWithoutMT/test/BaseWithoutMT.test.cpp
@@ -37,7 +37,7 @@ SCENARIO( "HeadRecord" ) {
 
     WHEN( "the data is given explicitly" ) {
 
-      double ZA = 1001;
+      int ZA = 1001;
       double AWR = 0.9991673;
 
       TestBaseWithoutMT chunk( ZA, AWR );
